[fix] Properly collapse jobs for most recent version
Originally the created_at timestamp was used for ordering the collapse
of the job records for a given job ID. There are some race conditions
that impact this which led to changes in how the jobs are
recorded (never in an async fashion). This still leaves the possibility
of creation dates being out of order but the records IDs will always be
properly sequential.